app/backend/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cryptography==36.0.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-migrate==4.00
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.5
PyMySQL==1.1.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.1
python-dotenv==1.0.0
debugpy==1.8.12
gunicorn==21.2.0
numpy==2.2.3
pyproj==3.7.1
mpmath==1.3.0
matplotlib==3.10.1